refactor: move global components to "app" dir
