+++ lib/app/app.js
@@ -1,15 +1,17 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Content from './Content'
-import OutboundLink from '../default-theme/OutboundLink.vue'
-import Badge from '../default-theme/Badge.vue'
-import ClientOnly from './ClientOnly'
 import dataMixin from './dataMixin'
 import store from './store'
 import { routes } from '@temp/routes'
 import { siteData } from '@temp/siteData'
 import enhanceApp from '@temp/enhanceApp'
 import themeEnhanceApp from '@temp/themeEnhanceApp'
+
+// built-in components
+import Content from './components/Content'
+import OutboundLink from './components/OutboundLink.vue'
+import Badge from './components/Badge.vue'
+import ClientOnly from './components/ClientOnly'
 
 // suggest dev server restart on base change
 if (module.hot) {
