+++ lib/app/components/OutboundLink.vue
